requirements.txt
flask==2.0.1
pytest==6.2.5
requests==2.26.0
gunicorn==20.1.0
psycopg2-binary==2.9.1